backend/requirements.txt
# FastAPI 核心
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# 数据库
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.10
alembic>=1.13.1

# 混合检索
rank-bm25>=0.2.2

# AI 模型（本地免费）
sentence-transformers>=2.3.1
torch>=2.0.0

# 文档处理
PyPDF2==3.0.1
pycryptodome>=3.20.0
python-docx==1.1.0
markdown==3.5.2

# 认证
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# 工具库
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-dotenv==1.0.0

# LLM 集成
openai>=1.12.0
httpx[socks]>=0.26.0
volcengine-python-sdk[ark]>=5.0.0

# 网络搜索
requests>=2.31.0
beautifulsoup4>=4.12.0

# 远程代码分析
paramiko>=3.4.0
anthropic>=0.40.0
